did several changes to be more platform independent
git-svn-id: file:///Volumes/Daten/wwwtech/cforum/branches/3.4@1481 31cb981c-0fcd-0310-a313-f9fa081e70fd

+++ src/clientlib.c
@@ -785,7 +785,7 @@ u_char *cf_advanced_get_link(const u_char *link,u_int64_t tid,u_int64_t mid,u_ch
 /* }}} */
 
 /* {{{ cf_general_get_time */
-u_char *cf_general_get_time(u_char *fmt,u_char *locale,int *len,time_t *date) {
+u_char *cf_general_get_time(u_char *fmt,u_char *locale,size_t *len,time_t *date) {
   u_char *buff;
   struct tm *tptr;
   size_t ln;
